requirement.txt
flask
tensorflow
argparse
Pillow
datetime
torch
opencv-python
opencv-python-headless
numpy
flask 
werkzeug
pandas
torchvision
tqdm
PyYAML
matplotlib
scipy
seaborn
jupyter

